% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBR_module.R
\name{CBR_module}
\alias{CBR_module}
\title{CBR Module}
\usage{
CBR_module(CBRCONF, n.folds = 10, KM.Plot = TRUE)
}
\arguments{
\item{CBRCONF}{A CBRCONF object}

\item{n.folds}{The number of folds}

\item{KM.Plot}{If KM.Plot is true, a kaplan-meier plot will be constructed from 3 risk groups (low, medium and high)}
}
\description{
This function performs n fold cross validation using the CBR-CONF framework. Confidence weights are used to determine
when retrieval ends for each test sample. Upon conclusion, the concordance index will be displayed.
}
