% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRCONF.R
\name{CBRCONF}
\alias{CBRCONF}
\title{CBRCONF Class}
\description{
CBRCONF Class

CBRCONF Class
}
\details{
This is a class object to store all of the results performed with the modules of this package
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{The dataset where time and status are the first and second columns and subsequent columns are features}

\item{\code{sig_feature_list}}{List of all significant features found in each iteration of feature selection}

\item{\code{sig_feature_data}}{Significant features that are retained after all iterations of feature selection}

\item{\code{feature_scores}}{A dataframe to contain all of the individual feature scores. This is the feature's expression value multiplied by the mean beta coefficient}

\item{\code{prognostic_scores}}{Dataframe of prognostic scores for all samples}

\item{\code{results_list}}{Holds the results for feature selection}

\item{\code{risk_groups}}{Holds the risk groups, which is assigned by the prognostic risk scores}

\item{\code{group_means}}{The features of each sample within each risk group is averaged to create one prototypical representation of each group. This stores that information}

\item{\code{low_risk_features}}{A subset of the original dataset that is just the low risk group, and their corresponding features}

\item{\code{medium_risk_features}}{A subset of the original dataset that is just the medium risk group, and their corresponding feature}

\item{\code{high_risk_features}}{A subset of the original dataset that is just the high risk group, and their corresponding feature}

\item{\code{low_risk_prototype}}{A vector of the averaged features of the low risk group}

\item{\code{medium_risk_prototype}}{A vector of the averaged features of the medium risk group}

\item{\code{high_risk_prototype}}{A vector of the averaged features of the high risk group}

\item{\code{confidence_weights_low}}{The confidence weights for just the low risk group}

\item{\code{confidence_weights_medium}}{The confidence weights for just the medium risk group}

\item{\code{confidence_weights_high}}{The confidence weights for just the high risk group}

\item{\code{confidence_weights}}{A dataframe that combines the confidence weights of the three risk groups}

\item{\code{data.folds}}{Each of the training and testing folds}

\item{\code{training_folds}}{Each of the training folds}

\item{\code{testing_folds}}{Each of the testing folds}

\item{\code{case_retrieval_information}}{Which cases were retrieved by which case}

\item{\code{testing_number_retrieved}}{How many cases each testing sample retrieved}

\item{\code{testing_concordance}}{The concordance values for each testing fold}

\item{\code{testing_sample_information}}{The risk group, predicted survival times, status, and testing_results}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CBRCONF$new()}}
\item \href{#method-clone}{\code{CBRCONF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new CBRContainer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CBRCONF$new(dataset)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new 'CBRMSR' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CBRCONF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
