% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_module.R
\name{cox_module}
\alias{cox_module}
\title{Cox Module
This module performs the feature selection step, in which it runs for Num iterations. At each iteration, a training set
and testing set will be created and every feature will run through a multivariate cox regression. After Num runs, features found to be
significant in the percent of training sets specified with Percent are kept (log ranking < 0.01 is used for significance). Cores
and blocksize are used for a parallel process (uses RegParallel). Cores is the number of computing cores assigned to the task, and
blocksize is the number of features to run at once. Blocksize should not exceed the number of features in your dataset.}
\usage{
cox_module(CBRCONF, Num, Percent, Cores)
}
\arguments{
\item{CBRCONF}{A CBRCONF object}

\item{Num}{The number of iterations to run}

\item{Percent}{Features recurring in this percent of training sets will be kept. Needs to be a decimal value between 0 and 1 (Ex. 50% is 0.5)}

\item{Cores}{The number of cores to use for parallel processing}
}
\description{
Cox Module
This module performs the feature selection step, in which it runs for Num iterations. At each iteration, a training set
and testing set will be created and every feature will run through a multivariate cox regression. After Num runs, features found to be
significant in the percent of training sets specified with Percent are kept (log ranking < 0.01 is used for significance). Cores
and blocksize are used for a parallel process (uses RegParallel). Cores is the number of computing cores assigned to the task, and
blocksize is the number of features to run at once. Blocksize should not exceed the number of features in your dataset.
}
