% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CBRCONF.R
\name{create_CBRCONF}
\alias{create_CBRCONF}
\title{Create a CBRCONF object}
\usage{
create_CBRCONF(dataset)
}
\arguments{
\item{dataset}{A dataset of features, where the first column is the time and the second column is the event statuses. Remaining columns are features.}
}
\description{
CBRCONF is an R6 class container that holds the dataset and relevant results at each stage.
}
