% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.R
\name{gbt}
\alias{gbt}
\title{Gradient Boosting Tree Module}
\usage{
gbt(dataset, n.folds = 10, n.tree = 500, blocksize = 1000, Cores = 7)
}
\arguments{
\item{dataset}{The dataset to use}

\item{n.folds}{The number of folds}

\item{blocksize}{The number of features to run through cox regression at once.}

\item{ntree}{The number of trees in the random survival forest}
}
\description{
This module will use gradient boosting tree to predict survival.
}
