% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSub.R
\docType{data}
\name{GeneSub}
\alias{GeneSub}
\title{Subset of gene expression data from the TCGA-BRCA dataset. This dataset is 500 random samples with 500 differentially expressed
genes found with TCGAbiolinks.}
\format{
An object of class \code{data.frame} with 500 rows and 502 columns.
}
\usage{
data(GeneSub)
}
\description{
Subset of gene expression data from the TCGA-BRCA dataset. This dataset is 500 random samples with 500 differentially expressed
genes found with TCGAbiolinks.
}
\examples{
data(GeneSub)
}
\keyword{datasets}
