% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_module.R
\name{knn_module}
\alias{knn_module}
\title{KNN Module}
\usage{
knn_module(CBRContainer, k = 5, n.folds = 10)
}
\arguments{
\item{CBRContainer}{A CBRContainer object}

\item{n.folds}{The number of folds}

\item{K}{The value of K for K Nearest Neighbors}

\item{KM.Plot}{If KM.Plot is true, a kaplan-meier plot will be constructed from 3 risk groups (low, medium and high)}
}
\description{
This module will use K-nearest neighbor to predict survival.
}
