% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoKNN.R, R/protoKNN_averagedistance.R
\name{protoKNN}
\alias{protoKNN}
\title{KNN Module}
\usage{
protoKNN(CBRContainer, n.folds = 10, KM.Plot = TRUE)

protoKNN(CBRContainer, n.folds = 10, KM.Plot = TRUE)
}
\arguments{
\item{CBRContainer}{A CBRContainer object}

\item{n.folds}{The number of folds}

\item{KM.Plot}{If KM.Plot is true, a kaplan-meier plot will be constructed from 3 risk groups (low, medium and high)}
}
\description{
This module will use K-nearest neighbor to predict survival.

This module will use K-nearest neighbor to predict survival.
}
