% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_distance_module.R
\name{prototype_distance_module}
\alias{prototype_distance_module}
\title{Prototype Distance Module}
\usage{
prototype_distance_module(CBRCONF)
}
\arguments{
\item{CBRCONF}{A CBRCONF object}
}
\description{
This module builds the distance matrices that determine how similar the samples are. Euclidean distance is always used.
Confidence weights will start to be built in this module (at this stage, only the distance from the prototype is calculated)
}
