% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_module.R
\name{prototype_module}
\alias{prototype_module}
\title{Prototype Module}
\usage{
prototype_module(CBRCONF)
}
\arguments{
\item{CBRCONF}{A CBRCONF object}
}
\description{
This module creates 3 prototypical representations of the risk groups. This is performed by using all samples within a risk
group and taking an average of each feature. Specifically, a vector is constructed for low, medium and high risk that has
averaged values for each feature.
}
