% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_groups_module.R
\name{risk_groups_module}
\alias{risk_groups_module}
\title{Risk Groups Module}
\usage{
risk_groups_module(CBRCONF, KM.Plot = TRUE)
}
\arguments{
\item{CBRCONF}{A CBRCONF object}

\item{KM.Plot}{If KM.Plot is true, a KM plot of 3 risk groups (low, medium, high) will be created and stored in the current work directory}
}
\description{
This module constructs 3 prognostic risk groups, taken from the ile function. In other words, each risk group is derived from
a tertile of the prognostic score. A KM plot can be constructed at this point from the actual survival data.
}
